+++ tests/test_queue.py
@@ -1,12 +1,23 @@
 import thoonk
 import unittest
+from ConfigParser import ConfigParser
+
 
 class TestQueue(unittest.TestCase):
 
     def __init__(self, *args, **kwargs):
         unittest.TestCase.__init__(self, *args, **kwargs)
-        self.ps = thoonk.Pubsub(db=10)
-        self.ps.redis.flushdb()
+
+        conf = ConfigParser()
+        conf.read('test.cfg')
+        if conf.sections() == ['Test']:
+            self.ps = thoonk.Thoonk(host=conf.get('Test', 'host'),
+                                    port=conf.getint('Test', 'port'),
+                                    db=conf.getint('Test', 'db'))
+            self.ps.redis.flushdb()
+        else:
+            print 'No test configuration found in test.cfg'
+            exit()
 
     def test_basic_queue(self):
         """Test basic QUEUE publish and retrieve."""
@@ -21,4 +32,4 @@ class TestQueue(unittest.TestCase):
         self.assertEqual(r, ["10", "20", "30", "40"], "Queue results did not match publish.")
 
 suite = unittest.TestLoader().loadTestsFromTestCase(TestQueue)
-            
+
